Always free items unconditionally
Items need to have their status tracked, in case they contain nested types that require freeing, to prevent double freeing. This is true, no matter where the tracking itself happens. Seeing that most objects will likely be plain-old-types for this project, make it the object's free function's responsibility to know when it shouldn't be executed multiple times.

Option would be to add metadata to each element in the pool and to do pointer arithmetic, when required.
(Could be determined by prescene of a free function for a type).